[Exception Replay] Mitigating an exception thrown while processing the methods participating in exception stack traces (#5783)
## Summary of changes
There was an exception that was thrown while processed the methods
participating in exception stack traces. the getter of
`MemberInfo.MetadataToken` threw `InvalidOperationException`. When it
happened, the whole processing of the exception was shuttered, leading
to missing debug info for these errors in Error Tracking.

## Reason for change
Make the processing of the exception more resilient to a broader
`MethodBase` derived implementations.

## Implementation details
Touching the `MetadataToken` property with a try/catch block, and
skipping the methods that yield any exception.