[Dynamic Instrumentation] Fixed Instrumentation failures revealed by the Exploration Tests of Line & Method probes (#5784)
## Summary of changes
In the past few weeks, we've been working internally to extend the
Exploration Tests of DI to work in conjunction with Line Probes,
alongside the current support of Method Probes. Meaning, for each
JIT-compiled method, we apply method probe instrumentation (surrounding
the method body with try/catch, unify the return points, capture method
variables on entry and leave, etc) with line instrumentation of each
line in the given method. By doing so, we test an extreme scenario of
instrumentation and validate our stability. By employing this technique,
we found a bunch of instrumentation issues that prevented Protobuf from
passing green with the aforementioned instrumentations. This PR
introduces fixes that makes Protobuf run to completion without any
instrumentation issues in its 1000+ tests.

## Reason for change
Make DI instrumentation more stable in extreme Method & Line probe
instrumentations.

## Implementation details
This PR introduced a fix for the following issues:
- When the line probe was applied to the beginning of the method, and
this beginning instruction is a target of a loop (when, for example, the
method body starts with a loop) - we failed to place the method probe
instrumentation in a way that will surround the line instrumentation
being applied to the first instruction.
- When a line probe instruction lands on a branch or a NOP instruction,
we skip it to fine-tune the selection. In the protobuf exploration test,
we encountered a scenario where the last instruction in the method body
was a branch / NOP. In which case, going one step further encountered
the first instruction (as the bytecode list is circular).
- When the line probe was placed on the end of a try block and the start
of the handler (catch or finally), we have to fix the branch targets
accordingly. We failed to do so in this specific scenario.


## Test coverage
Will be introduced in a separate PR that will enable the exploration
test of line & method.