[SSI] Bail out on known-faulty .NET 6 version (#5761)
## Summary of changes

- Bails out of SSI by default on .NET 6 < 6.0.12
- Tweak log message to write correct reason for the bail out

## Reason for change

There's a known crashing bug in early versions of .NET 6:
https://github.com/dotnet/runtime/pull/78670

We want to bail out of these cases by default in SSI. (Other major
versions are unaffected we believe)

## Implementation details

Check if we're in .NET 6 and build version is < 13, if so, bail out by
default.

## Test coverage

Manually tested locally to confirm we bail out in SSI if we're running
6.0.12, but not otherwise

## Other details


<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->