[Dynamic Instrumentation] Fixed SymDB upload when PDB is absent (#5789)
## Summary of changes
We had some issues in building up the symbols when PDB is absent which
lead to failure in uploading them.

## Reason for change
We are entering into the Open Beta of "Autocomplete and Search" feature
of the Dynamic Instrumentation product and want to make sure things are
working before shipping it. The absent of PDBs is a common scenario and
we still want to display Methods and Classes information that exist in
the assembly itself. Failing to build the symbols may yield bad user
experience and possibly hurt the adoption of DI for .NET.

## Implementation details
Mainly fixed a miscalculation of an index that lead to
`IndexOutOfBoundsException`.