Record when we use v2 instrumentation with a v3 version of the manual tracer (#5791)
## Summary of changes

- Adds an integration for v3 custom instrumentation to the v2 tracer
- Log a warning and metric when detected

## Reason for change

V3 custom instrumentation is not supported with v2 automatic
instrumentation. Currently, it just does _nothing_ which could be
confusing for the customer, as instrumentation is apparently working.
Additionally, we have no visibility of customers hitting this scenario.
By adding a log + metric we can identify how prevalent the issue is, and
give customers a reason/way forward for the behaviour.

## Implementation details

Add an integration that targets `Tracer.Instance` in v3+ of
`Datadog.Trace.Manual`. Record a log and a metric (once) if this is
called.

Alternative options would be to add something to Datadog.Trace.Manual
_itself_, but that's harder than it may sound initially because we don't
have (or want) any of the logging/metrics infra in that assembly, so I
think this is a good middle ground.

We will delete this integration in the v3 branch, as it is not required.

## Test coverage

Did a manual test locally and confirmed it behaves as expected. Given
we're going to remove this integration in v3, I don't think it's worth
spending time on automated testing of it.

## Other details

Once this is approved, I'll create a PR to update the metrics intake
with the `unsupported_custom_instrumentation_services` metric