[v3] Remove `DD_TRACE_OTEL_LEGACY_OPERATION_NAME_ENABLED` (#5787)
## Summary of changes

Removes the `DD_TRACE_OTEL_LEGACY_OPERATION_NAME_ENABLED` which when
enabled reverted the `OperationName` to be the
`ActivitySourceName.ActivityKind`. This was disabled by default.

## Reason for change

Unused and has since been replaced by a better `OperationName`
methodology by looking at various tags and span kind on the `Activity`.

## Implementation details

Removed the config key, removed the usage in `OtlpHelpers` and removed
the snapshots/tests that used it.

## Test coverage

- Deleted all snapshots and recreated them without the suffix and use of
the environment variable

## Other details
<!-- Fixes #{issue} -->

TODO Need to update the documentation to indicate that this was removed

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->