Package musl assets in linux glibc tar folder
To simplify onboarding (specifically with SSI), we package the musl
assets in the glibc tar file. Note that the _root_ of the tar
file still contains the glibc links, so nothing changes from a customer
perspective, but they _can_ link to the arch sub folder.

In v3, the files in the root become universal, and so the tar file can
truly be used the same way on glibc and musl. Moreover, in v3 we can
hardlink these to reduce the size. In v2, we can't do that, but there's
still value in doing it.

Note also that  we have left the `-musl.tar.gz` package with only the `-musl`
artifacts (i.e. we didn't package the glibc assets there). The reasoning
being that customers have already _explicitly_ selected musl at this
point, so the glibc assets are very unlikely to be used, and it doesn't
help us from an SSI perspective. Alpine customers are also likely more
size-conscious. We could potentially consider deprecating the artifact
 in a future major release, but don't plan to currently.