Build OCI images for every branch (#5800)
## Summary of changes

- Adds a convenience stage to Azure Devops for collating SSI artifacts
(tar.gz)
- Update OCI/deb/rpm to build based on tar artifacts instead of deb/rpm
- Add stages to GitLab to automatically build and test the OCI/deb/rpm
images on every branch

## Reason for change

- We want to switch to using the tar.gz artifacts as we're going to
include the musl artifacts in there later, so this makes things easier
- We currently are YOLOing the OCI releases without any testing. That's
bad. This adds some testing at least.

## Implementation details

Brace yourself for some contrived situations...

1. When a branch is pushed, GitLab triggers a build, which delays for
20mins...
2. Azure Devops build runs on PRs/master/manually etc.
3. Azure Devops creates an `ssi-artifacts` artifact that contains all
the tar.gz files (x64/musl-x64/arm64)
4. After 20 mins, the Gitlab job triggers, calls the Azure Devops REST
API to list the builds associated with the current branch, and searches
for the first build that matches the commit (yes, there are edge cases
here, no, we don't care too much right now tbh)
5. The GitLab job tries to grab the `ssi-artifacts` for the given build.
If it's not there, it sleeps and polls until it's available, or until it
times out (again, hacky, but YOLO)
6. If it finds the artifacts, it downloads them, and generates the SSI
build and test pipeline stages
7. Generate deb/rpm/OCI images based on the downloaded tar.gz artifacts
(when running in the branch) or the _release_ tar files (when running as
part of the release)
8. Runs the OCI/deb tests against the artifacts 🎉 

## Test coverage

- Ran a test
[here](https://gitlab.ddbuild.io/DataDog/dd-trace-dotnet/-/pipelines/39929160)
and all looks good AFAICT
- Compared the generated deb and OCI images, and they look the same to
me (except one small caveat, the `tar` file versions include a
`.scripts` folder that wasn't there in the debs, but that's benign.
- My only concern is that the `build-oci.sh` scripts break when we
_Actually_ do the release, but I think we'll deal with that as we come
to it. As long as the updated `curl` command is correct, I think we
should be pretty safe 🤞

## Other details

Caveats:

- Based on this PR by @robertomonteromiguel:
  - https://github.com/DataDog/dd-trace-dotnet/pull/5498
- This will all be replaced by one-pipeline soon, but we still want to
get this in ASAP
- We don't currently test the musl artifacts
- There's some annoying differences between the oci and deb scripts that
I'd love to normalize, but now is not the time...