[Exception Replay] Improved exceptions capturing accuracy + fixed a crash caused by mishandling of exception case probe statuses (#5798)
## Summary of changes
- Improved the accuracy of capturing exceptions.
- Improved the diagnostic capabilities for the scenarios we fail to
capture exception info, by introducing a new internal tag named
`_dd.di.er` that will hold the phase of the exception case.
- Fixed a crash caused by querying invalid probe ids for instrumentation
status via P/Invoke.

## Reason for change
There was a crash one of the early adopters of Exception Replay
experienced. We also got feedback that there are certain exceptions
which could not be captured by Exception Replay, thus showing no debug
information. This PR aims to improve that.

## Implementation details
- Prevented exception probes from querying instrumentation status from
the native side, similar safeguard is put in place in the native side as
well.
- Handling a race condition that prevented us from tracking an exception
to completion, when participating methods has failed in the lookup phase
for instrumentation.
- Unlocked the ability to put probe in inner classes beyond one level of
nesting. Previously, we supported only 1-level - now we support any
level of nesting. It improves the coverage of Dynamic Instrumentation
probes, and as result the coverage of Exception Replay.
- Improved our ability to proactively troubleshoot capturing issues of
Exception Replay by introducing a new inner tag named `_dd.di.er` which
holds a value representing the phase in which the exception resides. We
can see this tag from within our Support Admin to have a better
understating why a given case was not captured. Going forward, we may
choose to use it for a better UX by communicating the phase with the
customer.

## Test coverage
- Added a new test `RecursionWithInnerRefStructTest` that covers the
nesting improvement.
- Exception Replay testing suite is WIP.