Map `http.status_code` to meta dictionary (#5782)
## Summary of changes

This remaps the deprecated `http.status_code` tag to `meta` instead of
`metrics` for OpenTelemetry.

## Reason for change

This tag is supposed to be in `meta` but was going in `metrics`

## Implementation details

Same as #5654

## Test coverage

Added test to validate that it went to `metrics` and then correctly goes
to `meta`.
System tests also have this.

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->