Tracer flare - Inspect the AGENT_CONFIG content to set the log level (#5802)
## Summary of changes

When receiving the AGENT_CONFIG remote config, inspect the json to
extract the log level, instead of relying only on the filename.

## Reason for change

The filename is subject to change.

## Implementation details

In theory we should remove entirely the old behavior (checking the file
name) but the AGENT_CONFIG seems broken at the moment so I can't
properly test it. To play it safe, I just merged the new and old logic.

## Test coverage

Added a test case with a file name that doesn't contain the log level.