[Samples] Update IIS sample Dockerfile (#5805)
## Summary of changes

Improve the `Dockerfile` used to install `dd-trace-dotnet` in the
containerized IIS app sample in `tracer/samples/IISInDocker`.

## Reason for change

A user reported having issues with the Dockerfile, and I was able to
reproduce.

## Implementation details

- move install script to a separate `ps1` file
- use absolute paths for msi download and log file (this is the main
reason for this PR, the relative paths fail in common scenarios)
- use PowerShell's built-in `Invoke-WebRequest` instead of .NET's
`System.Net.WebClient.DownloadFile()`
(which is
[obsolete](https://learn.microsoft.com/en-us/dotnet/api/system.net.webclient#remarks))
- get the github release flagged as `latest` instead of the most recent
(not the same thing! this would've downloaded preview versions of v3.x,
for example)
- add more verbosity in console messages (show url, download path, log
path, etc) to ease troubleshooting
- add basic `.dockerignore` file

## Test coverage

Tested manually.

<!--
## Other details
-->

<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->