Add Callsite aspects analyzer to check for "safe" patterns (#5835)
## Summary of changes

Adds two analyzers that check for "safe" usage of the callsite arguments

## Reason for change

In CallTarget we explicitly wrap all the custom code in try-catch
statements for safety. It's not technically easy to do that with
CallSite, so this provides the next best thing - an analyzer that
enforces specific patterns.

## Implementation details

There are two analyzers included

- `BeforeAfterAspectAnalyzer` that works with
`[AspectMethodInsertBefore]` and `[AspectMethodInsertAfter]`
- `ReplaceAspectAnalyzer` that works with `[AspectCtorReplace]` and
`[AspectMethodReplace]`

> The CodeFixProvider for `ReplaceAspectAnalyzer` is "best effort", and
only works for aspects that follow an expected pattern, as otherwise
there's too much risk of introducing a bug.

The easiest way to show what they do is to look at the unit tests, or
the following GIFs of the analyzers and code fixes in action

![aspect
analyzer](https://github.com/user-attachments/assets/6b09c721-8d6f-4e37-bfba-dc60e9a6ad99)

![aspect
analyzer2](https://github.com/user-attachments/assets/b796c6ce-f88e-4a77-a34c-75336427318f)

## Test coverage

Unit tests for the behaviour. Also verified that they show up in the IDE

## Other details

Setting the severity of these analyzers to `Warning` or `Error` breaks
the build, as most of our aspects are non-compliant. For now, I've set
it to `Info` so that nothing breaks, but you only get a minimal info
about the error in the IDE:


![image](https://github.com/user-attachments/assets/4d343d12-a010-401c-a74b-1747d538fd5b)

When we fix these issues, we should be sure to change the severity of
the analyzers to `Error` first, as it will make them easier to fix, and
we will know we caught them all.