[ASM] Add RASP timeout flag (#5827)
## Summary of changes

Add the RASP timeout metric that is set if the WAF returns a timeout.

The tag is defined
[here](https://docs.google.com/document/d/1122UKiuvJao6cV5D_NqoExTL5GQjDpbGTJENkBbEgVc/edit).

__dd.appsec.rasp.timeout: should be included as a metric within the root
span of each request in which the WAF has had a timeout. This value must
be set to 1 if there has been a timeout, but may be omitted if there
hasn’t been one; alternatively it can also be set to 0._

## Reason for change

It helps to find problematic traces and detect timeout problems.

## Implementation details

## Test coverage

A unit test has been added.

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->

---------

Co-authored-by: Andrew Lock <andrew.lock@datadoghq.com>