[Dynamic Instrumentation] Improved instrumentation matching of symbols received through SymDb (#5829)
## Summary of changes
We are on the edge of advancing SymDb to Open Beta. We found a few
issues that blocked us in doing so:
- Generic types, methods and arguments received surrounded with `<` and
`>` while the native matcher is based on `[` and `]`.
- The parser of the signature in the managed side of the debugger
implemented naively, mainly splitted by `,`. Generic params are comma
separated too, which failed to parse it correctly.
- Static methods were not handled correctly in the native side. It
naively skipped the first param, as if all the methods it deals with are
instance methods.
- The signature received by SymDb is surrounded by `(` and `)` and has
space in between every param type.

## Reason for change
SymDb readiness for Open Beta.

## Implementation details
- In the managed side, upon receiving a request to instrument a method
with a signature that is targeting a generic method/type, the `<` and
`>` characters are replaced with `[` and `]` respectively.
- Parsing the signature in a more robust way that handles the generic
case where comma could present in between every generic param.
- Static methods are now correctly handled in the native side when a
request arrives with a signature exact matching.
- Correctly parsing the signature when it comes surrounded by `(` and
`)`, including space in between every param type.

## Test coverage
-
[SignatureParserTests](https://github.com/DataDog/dd-trace-dotnet/blob/a8a2df542625b29e3f0d2ce8c745717f27abdb0a/tracer/test/Datadog.Trace.Debugger.IntegrationTests/SignatureParserTests.cs)