Ensure top-level entry points are wrapped with try-catch (#5838 -> v2) (#5850)
- Ensure all "entry points" are guaranteed safe with try-catch
- Ensure all `Task.Run` observe exceptions

We want to make sure (as best we can) that we don't crash when customer
applications are instrumented. This is about the best we can do I think.

`UnobservedTaskException` isn't really a _thing_ in .NET Core AFAICT,
and in .NET Framework it requires customers [explicitly turning it
on](https://learn.microsoft.com/en-us/dotnet/api/system.threading.tasks.taskscheduler.unobservedtaskexception?view=net-8.0)
using

```xml
<configuration>
   <runtime>
      <ThrowUnobservedTaskExceptions enabled="true"/>
   </runtime>
</configuration>
```

but IMO it's still preferable to ensure we wouldn't cause crashes in
these cases if we can help it

- Wraps "entry points" with a try-catch. These are invoked "directly"
due to how they're injected or called. e.g.
  - `Datadog.Trace.ClrProfiler.Managed.Loader.Startup.Startup()`
  - `AgentProcessManager.Initialize()`
  - `Instrumentation.Initialize()`
  - (`DiagnosticManager` is already handled)
- Checked places we use `Task.Run` or `TaskFactory.StartNew` and checked
that we're either:
  - Wrapping the called code in `try-catch`
  - Calling `await` on the returned `Task`
  - Using `ContinueWith` with at least `OnlyOnFaulted`

This is all defensive, so just checking that it hasn't introduced
regressions (covered by existing tests)

As Kevin pointed out, basically _all_ code can throw, so we can't
necessarily wrap _everything_, but hopefully this captures _most_ of the
"top level" cases that could cause "bubble up" issues for customers

Co-authored-by: Kevin Gosse <kevin.gosse@datadoghq.com>
Co-authored-by: Lucas Pimentel <lucas.pimentel@datadoghq.com>