[ASM] IAST Email HTML Injection vulnerability (#5780)
## Summary of changes

This PR contains the implementation of the email HTML injection
vulnerability.

Email HTML injection is a vulnerability where user input is included in
the content of an email without proper validation and sanitization. This
vulnerability can have severe consequences as it opens the door for
various attacks, including phishing, social engineering exploits, and
the exploitation of email client vulnerabilities.

This vulnerability is described and discussed in the following
[RFC](https://docs.google.com/document/d/1-_cAoSo3zUxlBNAp1f4M_4NRq6zdVzwtVyW_GaU-xGs/edit?pli=1#heading=h.tonjsgarlieo).

The evidence redaction was modified in all the libraries to redact
emails addresses, last names and user names, which affected some of our
previous tests. These changes were requested to cover some cases that
this new vulnerability is facing.

## Reason for change

The implementation for this vulnerability was approved for Q3.

## Implementation details

## Test coverage

Integration and instrumented unit tests were added.

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->