Skip the mass transit test to see if it solves flake issues (#5861)
## Summary of changes

Skip the mass transit smoke test as it seems to be a cause of a lot of
flakiness

## Reason for change

We've seen a lot of errors in the `CheckBuildlogsForErr` stage:

```
CheckBuildLogsForErr: 03:08:39 [Error] An error occurred while sending data to the agent at http://127.0.0.1:39573/v0.4/traces. If the error isn't transient, please check https://docs.datadoghq.com/tracing/troubleshooting/connection_errors/?code-lang=dotnet for guidance. System.Net.Http.HttpRequestException: Error while copying content to a stream.
```

These seemed to get a lot worse after we disabled keep-alive, but that's
anecdotal.

## Implementation details

It's not entirely clear if the problem is just coincidentally related to
the MassTransit test (i.e. it's a test ordering process) or if it's
actually something about the test.

As a check I tried skipping the test in this branch and did 4 full (all
TFM) integration tests runs, and didn't see the issue again. It's all
still anecdotal, but rather trade off flakiness here. If the problem
reappears subsequently, we can look into it again further.

## Test coverage

Did 4 full runs, and didn't see the issue again