Config refactor - Add telemetry to otel config (#5717 -> v2) (#5864)
## Summary of changes

Records logs and metrics for when otel config is overriden by datadog
config, or when it's invalid

## Reason for change

This was originally part of
https://github.com/DataDog/dd-trace-dotnet/pull/5661, but wasn't
possible due to concerns with the fact `ConfigurationBuilder` is used in
critical "startup" paths for the tracer, so can result in recursion if
we're not careful.

## Implementation details

The crux of the implementation is that we "store" error and metrics for
writing at the point we know it's safe. This is _similar_ to what we
already do for configuration telemetry.

(Technically I think we _Could_ directly access the
`TelemetryFactory.Metrics`, but this is "safer", and makes it easier to
test we're sending the right metrics

## Test coverage

Added unit tests to all of the existing settings tests where we override
configuration to confirm that it works as expected

## Other details

The metrics etc are all ported from
https://github.com/DataDog/dd-trace-dotnet/pull/5661, but there are
currently some issues with the proposed values:

- [x] The metrics aren't added to the intake yet, so need updating there
(and copying here)
- [x] The tag names aren't "valid" for standard telemetry metrics, they
should be lowercase, and shouldn't use `.`. I would suggest changing
these to be lowercase and replacing `.` with `_`, but either way this
should happen after they've been OK'd in the intake.

Original PR:
- https://github.com/DataDog/dd-trace-dotnet/pull/5717