Catch exceptions when trying to shutdown IIS (#5840)
## Summary of changes

Catch exceptions when sending a shutdown request to IIS.

## Reason for change

To shutdown IIS, we first send a request to the web app to ask it to
exit gracefully, and then we kill the IIS Express process. But if for
some reason an exception is thrown when sending the request, we wouldn't
get to the killing part.