Backport unified pipeline for APM SDKs for SSI artifacts (#5868)
## Summary of changes

This is a backport of #5818 to the 2.x release branch. The main
difference is that 2.x needs the init images for musl published under
the `-musl` tag suffix

## Reason for change

By using the unified pipeline across all APM libraries, there is less
maintenance hassle, fewer mistakes, and a more uniform creation of
artifacts.

## Implementation details

I kept all of the changes from the original PR isolated to a single
commit for ease of review. The goal of the changes was to do the minimal
without having to change `one-pipeline` upstream.

**download-single-step-artifacts.sh** : 
download the musl tarball from Github releases

**prepare-oci-package.sh** : 
if `LIBC` is `musl` use the `musl` tarball

**gitlab-ci.yml** :
* package-oci-musl : I couldn't change the `package-oci` matrix without
having dependency/artifact issues so a new job is created to build the
musl oci package. This package is only used to create the lib injection
image and the OCI package is not published anywhere
* create-arch-specific-lib-injection-image-musl: build the lib injection
image from package-oci-musl. The hack is that `ARCH` is abused to get
the tag to be `COMMIT_SHA-musl` . A multiach image is unneeded because
only amd64 supports musl and the single arch image is tagged correctly.
* generate-lib-init-pinned-tag-values : The `master` version overrode
one-pipeline to add `musl` tag to everything. That override is removed
* generate-lib-init-pinned-tag-values-musl: Generates the musl tags
* publish-lib-init-pinned-tags-musl : publishes the musl tags using
`COMMIT_SHA-musl` as the source

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->