Add an `IsManualInstrumentationOnly` flag to Datadog.Trace.Manual (#5866)
## Summary of changes

Adds a `IsManualInstrumentationOnly()` method to the manual library, and
rewrites it when we're doing automatic instrumentation

## Reason for change

For some scenarios we're considering (e.g. v3 part deux) we need to know
from inside the manual library whether we're auto-instrumenting. This is
a relatively simple approach to achieve it.

## Implementation details

Mostly copied C++ rewriting code that we're already doing, and tweaked
it for our use case

## Test coverage

Added a check to the manualinstrumentation tests that the flag behaves
as expected (`true` when no auto instrumentation, `false` if we have
auto instrumentation)

## Other details

For part deux we'll need to do more rewriting, but there shouldn't be
any harm in doing the rewriting in this PR regardless of what happens
with that