[ASM] ensure struct is on the stack before passing to native code (#5882)
## Summary of changes

Some crashes where reported where the stack trace showed the error
occurred in the WAF, with the call starting in
`Waf.UpdateWafAndDispose`. A review of this method spotted that a
pointer to a struct stored in a object field was being passed to the
WAF. This meant the GC might move the object, which could potentially
lead to a crash.