Add verification step to create_draft_release to check SSI one-pipeline succeeded (#5865 -> v2) (#5879)
## Summary of changes

Adds a stage to `create_draft_release` that verifies the gitlab stages
all passed before starting the release

## Reason for change

Before we merged #5818 we were testing the lib injection images prior to
publishing, but that PR removed the checks (we're building/testing
everything in the one pipeline instead)

This PR, _explicitly_ checks that everything passed in GitLab before
starting the release.

## Implementation details

Pulls all the statuses for the commit, and makes sure the GitLab ones
all have a passing status.

Pros: 
- If/when the one-pipeline expands, we automatically check for success.
- One pipeline can change (add/remove checks) and we automatically make
sure they all passed before doing a release.

Cons: 
- The one pipeline could break at some point without being a _real_
issue, which would block our releases. So added an override we can use
in case of emergency. It should only be used when the reporting is
"false" though, because the one pipeline creates and publishes the SSI
artifacts.
- We don't explicitly require any of the checks. This gives us
flexibility (they can add or remove checks) but we won't know if we're
actually checking them all (for example)
- If the GitHub/GitLab link isn't working (due to issues either side),
we might not be notified about failures

## Test coverage

Tested locally to confirm the `VerifyReleaseReadiness` works as
expected. e.g. run

```powershell
.\tracer\build.ps1 VerifyReleaseReadiness -CommitSha 89bf7b321e4b0643e62a54c9b3497fc19ac3615d -GITHUB_TOKEN <your token>
```

and it passes, but `ec735117ceaa963d5033f83ed80daaa88f970867` fails (for
example)

## Other details

Backport of #5865

---------

Co-authored-by: Kevin Gosse <kevin.gosse@datadoghq.com>